Added the silent command to the autoformat command. Caused error on :w after undo.
The error that occured when :w after undoing an action was:

Error detected while processing BufWritePre Auto commands for "*":
E790: undojoin is not allowed after undo

The issue is quite clear from the error why it occured. I use undojoin because i don't
want the formatter to be to noisy in the undo history. So it had to be
left.

The solution is a workaround, i call silent! before undojoin. This will
have the unfortunate effect that i wont get any errors if the formatter
errors out, but that's a rare issue. So i hope that this will work out
in the future.